Approach ======== We cannot make pair of (1,2).. otherwise sum will be less and we have make pair of type (2,1) to increase the sum.
Target --> Is to maximize (2, 1) and minimize (1, 2) pair

Now pairs (1,0), (0,1), (2,0) and (0,2) will not contribute any to the sum.

So first reduce those..

But it would be wise to reduce pair (1, 0) and (0, 2)
because of it will reduce number of 1 in seq1 and number of 2
in seq, Thus reducing 1,2 pairs

so let's say array is x0, x1, x2 and y0, y1, y2

So number of (1, 0) pair is

min(x1, y0) -->a1

x1 - a1

y0 - a1

Same for (0,2)

min(x0, y2) --> a2

x0 - a2
y2 - a2

Now we got (1,2) (minimized by above operation) and (2, 1)

Number of (2, 1)

min(x2, y1) --> a3

sum = 2 * a3

and min (x1, y2) --> a4

sum = - 2 * a4

sum is our ans. :)